Code cleanup after merges
