+++ lib/views/video_view/chat_video_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:gocast_mobile/models/user/mock_data.dart';
+import 'package:gocast_mobile/base/helpers/mock_data.dart';
 
 ///This class is used to display the chat view in the video player page.
 ///it is kind of temporary solution, we will replace it with the real chat view
